refactor contact us
