+++ src/components/ContactUs/ManagingDirectors.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { DynamicTrans } from '../utils';
+
+export const ManagingDirectors = () => (
+  <div className="col-6 p-0 text-nowrap">
+    <h6>
+      <DynamicTrans>Managing directors</DynamicTrans>
+    </h6>
+    <p>
+      Yoko Spirig
+      <br />
+      Ben-Elias Brandt
+      <br />
+      Timo Horstschäfer
+    </p>
+  </div>
+);
